+++ prog/probe.go
@@ -88,7 +88,6 @@ func probeMain(flags probeFlags) {
 		hostID   = hostName // TODO(pb): we should sanitize the hostname
 	)
 	log.Infof("probe starting, version %s, ID %s", version, probeID)
-	log.Infof("command line: %v", os.Args)
 	checkpointFlags := map[string]string{}
 	if flags.kubernetesEnabled {
 		checkpointFlags["kubernetes_enabled"] = "true"
@@ -170,7 +169,7 @@ func probeMain(flags probeFlags) {
 	}
 
 	if flags.kubernetesEnabled {
-		if client, err := kubernetes.NewClient(flags.kubernetesAPI, flags.kubernetesInterval); err == nil {
+		if client, err := kubernetes.NewClient(flags.kubernetesConfig); err == nil {
 			defer client.Stop()
 			reporter := kubernetes.NewReporter(client, clients, probeID, hostID, p, handlerRegistry)
 			defer reporter.Stop()
@@ -178,7 +177,7 @@ func probeMain(flags probeFlags) {
 			p.AddTagger(reporter)
 		} else {
 			log.Errorf("Kubernetes: failed to start client: %v", err)
-			log.Errorf("Kubernetes: make sure to run Scope inside a POD with a service account or provide a valid kubernetes.api url")
+			log.Errorf("Kubernetes: make sure to run Scope inside a POD with a service account or provide valid probe.kubernetes.* flags")
 		}
 	}